Rework according to feedback
